# openapi/pufferlab-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "enum": [
          "validation_error",
          "not_found",
          "namespace_not_ready",
          "provider_error",
          "rate_limited",
          "run_conflict",
          "internal_error"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Output"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "trace_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
...
      "CancelEvalRunResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "result": {
            "$ref": "#/components/schemas/EvalRunView"
          }
        },
        "required": [
          "result"
        ],
        "title": "CancelEvalRunResponse",
        "type": "object"
      },
...
      "ContractVersion": {
        "const": 1,
        "type": "integer"
      },
      "CreateEvalRunRequest": {
        "additionalProperties": false,
        "properties": {
          "baseline_config_id": {
            "format": "uuid",
            "title": "Baseline Config Id",
            "type": "string"
          },
          "candidate_config_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Candidate Config Ids",
            "type": "array"
          },
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "max_concurrency": {
            "default": 4,
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Max Concurrency",
            "type": "integer"
          },
          "query_set_id": {
            "format": "uuid",
            "title": "Query Set Id",
            "type": "string"
          },
          "random_seed": {
            "default": 20260822,
            "title": "Random Seed",
            "type": "integer"
          },
          "warmup_query_count": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Warmup Query Count",
            "type": "integer"
          }
        },
        "required": [
          "query_set_id",
          "baseline_config_id",
          "candidate_config_ids"
        ],
        "title": "CreateEvalRunRequest",
        "type": "object"
      },
      "CreateEvalRunResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "result": {
            "$ref": "#/components/schemas/EvalRunView"
          }
        },
        "required": [
          "result"
        ],
        "title": "CreateEvalRunResponse",
        "type": "object"
      },
...
      "EvalRunDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "result": {
            "$ref": "#/components/schemas/EvalRunView"
          }
        },
        "required": [
          "result"
        ],
        "title": "EvalRunDetailResponse",
        "type": "object"
      },
...
      "EvalRunExportResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "export": {
            "$ref": "#/components/schemas/EvalRunExport"
          }
        },
        "required": [
          "data_origin",
          "export"
        ],
        "title": "EvalRunExportResponse",
        "type": "object"
      },
      "EvalRunListResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/EvalRunView"
            },
            "maxItems": 100,
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "title": "EvalRunListResponse",
        "type": "object"
      },
      "EvalRunQueryDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "$ref": "#/components/schemas/DatasetAttribution"
          },
          "baseline_config_id": {
            "format": "uuid",
            "title": "Baseline Config Id",
            "type": "string"
          },
          "candidate_config_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Candidate Config Ids",
            "type": "array"
          },
          "configs": {
            "items": {
              "$ref": "#/components/schemas/RetrievalConfigSummary"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Configs",
            "type": "array"
          },
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "live_replay_policy_permitted": {
            "title": "Live Replay Policy Permitted",
            "type": "boolean"
          },
          "original_stage_evidence_available": {
            "const": false,
            "default": false,
            "title": "Original Stage Evidence Available",
            "type": "boolean"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/EvalOutcomeRecord"
            },
            "maxItems": 4,
            "title": "Outcomes",
            "type": "array"
          },
          "query": {
            "$ref": "#/components/schemas/JudgedQuery"
          },
          "rank_changes": {
            "items": {
              "$ref": "#/components/schemas/CandidateRelevantRankChanges"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Rank Changes",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "data_origin",
          "query",
          "baseline_config_id",
          "candidate_config_ids",
          "configs",
          "outcomes",
          "rank_changes",
          "attribution",
          "live_replay_policy_permitted"
        ],
        "title": "EvalRunQueryDetailResponse",
        "type": "object"
      },
      "EvalRunQueryReplayRequest": {
        "additionalProperties": false,
        "properties": {
          "config_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Config Ids",
            "type": "array"
          },
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "include_counterfactual_probe": {
            "default": false,
            "title": "Include Counterfactual Probe",
            "type": "boolean"
          }
        },
        "required": [
          "config_ids"
        ],
        "title": "EvalRunQueryReplayRequest",
        "type": "object"
      },
...
      "RegressionOrder": {
        "enum": [
          "regressions",
          "gains"
        ],
        "title": "RegressionOrder",
        "type": "string"
      },
...
      "RegressionResponse": {
        "additionalProperties": false,
        "properties": {
          "baseline_config_id": {
            "format": "uuid",
            "title": "Baseline Config Id",
            "type": "string"
          },
          "candidate_config_id": {
            "format": "uuid",
            "title": "Candidate Config Id",
            "type": "string"
          },
          "contract_version": {
            "$ref": "#/components/schemas/ContractVersion",
            "default": 1
          },
          "coverage": {
            "$ref": "#/components/schemas/RegressionCoverage"
          },
          "data_origin": {
            "$ref": "#/components/schemas/DataOrigin"
          },
          "limit": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "order": {
            "$ref": "#/components/schemas/RegressionOrder"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RegressionRow"
            },
            "maxItems": 50,
            "title": "Rows",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "data_origin",
          "baseline_config_id",
          "candidate_config_id",
          "order",
          "limit",
          "rows",
          "coverage"
        ],
        "title": "RegressionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Search evaluation and query-forensics workbench",
    "title": "PufferLab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/eval-runs": {
      "get": {
        "operationId": "list_evaluation_runs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "List Evaluation Runs",
        "tags": [
          "evaluation runs"
        ]
      },
      "post": {
        "operationId": "create_evaluation_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvalRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEvalRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Create Evaluation Run",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}": {
      "get": {
        "operationId": "get_evaluation_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Get Evaluation Run",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_evaluation_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelEvalRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Cancel Evaluation Run",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}/export": {
      "get": {
        "operationId": "export_evaluation_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Export Evaluation Run",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}/queries/{query_id}": {
      "get": {
        "operationId": "get_evaluation_run_query",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunQueryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Get Evaluation Run Query",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}/queries/{query_id}/replay": {
      "post": {
        "operationId": "replay_evaluation_run_query",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRunQueryReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunQueryReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Replay Evaluation Run Query",
        "tags": [
          "evaluation runs"
        ]
      }
    },
    "/api/v1/eval-runs/{run_id}/regressions": {
      "get": {
        "operationId": "get_evaluation_regressions",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidate_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Config Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RegressionOrder",
              "default": "regressions"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation record was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The requested evaluation action conflicts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The evaluation request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "The request failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                }
              }
            },
            "description": "Evaluation data or runtime is unavailable."
          }
        },
        "summary": "Get Evaluation Regressions",
        "tags": [
          "evaluation runs"
        ]
      }
    },
...
  }
}